# openapi/openapi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "PagoPA API - Reporting Organizations Enrollments",
		"description": "Reporting orgs enrollment project",
		"termsOfService": "https://www.pagopa.gov.it/",
		"version": "0.0.1"
	},
	"servers": [
		{
			"url": "http://localhost:8080",
			"description": "Generated server url"
		}
	],
	"paths": {
		"/info": {
			"get": {
				"tags": [
					"Home"
				],
				"summary": "health check",
				"description": "Return OK if application is started",
				"operationId": "healthCheck",
				"responses": {
					"200": {
						"description": "OK",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AppInfo"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"429": {
						"description": "Too many requests",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"500": {
						"description": "Service unavailable",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					}
				},
				"security": [
					{
						"ApiKey": []
					},
					{
						"Authorization": []
					}
				]
			},
			"parameters": [
				{
					"name": "X-Request-Id",
					"in": "header",
					"description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
					"schema": {
						"type": "string"
					}
				},
				{
					"name": "X-Request-Id",
					"in": "header",
					"description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
					"schema": {
						"type": "string"
					}
				}
			]
		},
		"/organizations": {
			"get": {
				"tags": [
					"Reporting Organizations Enrollments API"
				],
				"summary": "The organization get all enrollments of creditor institutions.",
				"operationId": "getOrganizations",
				"responses": {
					"200": {
						"description": "Obtained all enrollments.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/OrganizationModelResponse"
									}
								}
							}
						}
					},
					"401": {
						"description": "Wrong or missing function key.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"500": {
						"description": "Service unavailable.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					}
				},
				"security": [
					{
						"ApiKey": []
					},
					{
						"Authorization": []
					}
				]
			},
			"parameters": [
				{
					"name": "X-Request-Id",
					"in": "header",
					"description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
					"schema": {
						"type": "string"
					}
				},
				{
					"name": "X-Request-Id",
					"in": "header",
					"description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
					"schema": {
						"type": "string"
					}
				}
			]
		},
		"/organizations/{organizationFiscalCode}": {
			"get": {
				"tags": [
					"Reporting Organizations Enrollments API"
				],
				"summary": "The organization get the single enrollment for the creditor institution.",
				"operationId": "getOrganization",
				"parameters": [
					{
						"name": "organizationFiscalCode",
						"in": "path",
						"description": "The fiscal code of the Organization.",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Obtained single enrollment.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrganizationModelResponse"
								}
							}
						}
					},
					"401": {
						"description": "Wrong or missing function key.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"404": {
						"description": "Not found the enroll service.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					},
					"500": {
						"description": "Service unavailable.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					}
				},
				"security": [
					{
						"ApiKey": []
					},
					{
						"Authorization": []
					}
				]
			},
			"post": {
				"tags": [
					"Reporting Organizations Enrollments API"
				],
				"summary": "The organization creates a creditor institution.",
				"operationId": "createOrganization",
				"parameters": [
					{
						"name": "organizationFiscalCode",
						"in": "path",
						"description": "The fiscal code of the Organization.",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Request created.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrganizationModelResponse"
								}
							}
						}
					},
					"401": {
						"description": "Wrong or missing function key.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"409": {
						"description": "The organization to create already exists.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					},
					"500": {
						"description": "Service unavailable.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					}
				},
				"security": [
					{
						"ApiKey": []
					},
					{
						"Authorization": []
					}
				]
			},
			"delete": {
				"tags": [
					"Reporting Organizations Enrollments API"
				],
				"summary": "The organization deletes the creditor institution.",
				"operationId": "removeOrganization",
				"parameters": [
					{
						"name": "organizationFiscalCode",
						"in": "path",
						"description": "The fiscal code of the Organization.",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Request deleted.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"401": {
						"description": "Wrong or missing function key.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"404": {
						"description": "Not found the creditor institution.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					},
					"500": {
						"description": "Service unavailable.",
						"headers": {
							"X-Request-Id": {
								"description": "This header identifies the call",
								"schema": {
									"type": "string"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemJson"
								}
							}
						}
					}
				},
				"security": [
					{
						"ApiKey": []
					},
					{
						"Authorization": []
					}
				]
			},
			"parameters": [
				{
					"name": "X-Request-Id",
					"in": "header",
					"description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
					"schema": {
						"type": "string"
					}
				},
				{
					"name": "X-Request-Id",
					"in": "header",
					"description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
					"schema": {
						"type": "string"
					}
				}
			]
		}
	},
	"components": {
		"schemas": {
			"ProblemJson": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string",
						"description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable"
					},
					"status": {
						"maximum": 600,
						"minimum": 100,
						"type": "integer",
						"description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
						"format": "int32",
						"example": 200
					},
					"detail": {
						"type": "string",
						"description": "A human readable explanation specific to this occurrence of the problem.",
						"example": "There was an error processing the request"
					}
				}
			},
			"OrganizationModelResponse": {
				"type": "object",
				"properties": {
					"organizationFiscalCode": {
						"type": "string"
					},
					"organizationOnboardingDate": {
						"type": "string"
					}
				}
			},
			"AppInfo": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					},
					"version": {
						"type": "string"
					},
					"environment": {
						"type": "string"
					}
				}
			}
		},
		"securitySchemes": {
			"ApiKey": {
				"type": "apiKey",
				"description": "The API key to access this function app.",
				"name": "Ocp-Apim-Subscription-Key",
				"in": "header"
			}
		}
	}
}